requirements.txt

Django
djangorestframework
djangorestframework-simplejwt
django-filter
drf-yasg
python-dotenv
